+++ include/mem/simd_scanner.h
@@ -110,7 +110,7 @@ namespace mem
                     } while (true);
 
                     ++current;
-                    current = find_byte(current + skip_pos, pat_bytes[skip_pos], end - current) - skip_pos;
+                    current = find_byte(current + skip_pos, pat_bytes[skip_pos], static_cast<std::size_t>(end - current)) - skip_pos;
                 }
 
                 return nullptr;
@@ -133,7 +133,7 @@ namespace mem
                     } while (true);
 
                     ++current;
-                    current = find_byte(current + skip_pos, pat_bytes[skip_pos], end - current) - skip_pos;
+                    current = find_byte(current + skip_pos, pat_bytes[skip_pos], static_cast<std::size_t>(end - current)) - skip_pos;
                 }
 
                 return nullptr;
